move db version handlers to version.go
