+++ node-registrar/pkg/db/nodes.go
@@ -117,33 +117,3 @@ func (db *Database) CreateUptimeReport(report *UptimeReport) error {
 		return nil
 	})
 }
-
-func (db *Database) SetZOSVersion(version string) error {
-	var current ZosVersion
-	result := db.gormDB.Where(ZosVersion{Key: ZOS4VersionKey}).Attrs(ZosVersion{Version: version}).FirstOrCreate(&current)
-
-	if result.Error != nil {
-		return result.Error
-	}
-
-	if result.RowsAffected == 0 {
-		if current.Version == version {
-			return errors.New("version already set")
-		}
-		return db.gormDB.Model(&current).
-			Select("version").
-			Update("version", version).Error
-	}
-	return nil
-}
-
-func (db *Database) GetZOSVersion() (string, error) {
-	var setting ZosVersion
-	if err := db.gormDB.Where("key = ?", "zos_4").First(&setting).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return "", ErrRecordNotFound
-		}
-		return "", err
-	}
-	return setting.Version, nil
-}
